dcp-815 separate client builder
